feature: seperate apis to adopt microservices architecture
